chore: Move most of the test suites to PHPUnit 10
Not bumping directly to PHPUnit 11, we will have to bump to
Psalm 6 first due to conflicts with a transitive dep. This is
also temporarily force us to disable the coverage collection in
the nightly pipeline.

Calls to ::withConsecutive have been replaced using Rector
https://getrector.com/rule-detail/with-consecutive-rector
Psalm baseline has been re-generated to include the new errors
from this automated migrations.

Our custom extension related to Mockery has been removed in order to not
have to port it to the new system. Given we are not introducing new
Mockery usages it should not be an issue.

There are still a bunch of PHPUnit deprecations, they will be dealt with
in a independent PR, this change is already large enough.

Part of request #41810 PHPUnit: 9 -> 11

Change-Id: I2d639e3ac1b6a7d7829bf6651c3f09473378da40