refactor: move link field to dedicated lib
part of story #41466 Have new Artifact Links in Artifact View

No functional change expected.

Why?
The goal of this refactoring is to extract the Artifact Links field from
the Artifact modal to its own library, so that we can reuse it in the
full-page Artifact view. The link field and the modal itself now depend
on the "artifact-common" library for anything shared between them, such
as Identifier types and events.

I also took the opportunity to simplify some identifier creations
(UserIdentifier and TrackerShortname) by "just" having a constructor
from a number or string, instead of a Proxy + interface. This is an
application of ADR-0005 in Program Management [0], which is also valid
in the link field as some concepts have been used here too.

We cannot use async gettext because the link-field is used by the modal,
which uses webpack, and it does not seem to pick up the dynamic imports
to re-export them as Vite does in Artidoc. It's ok as we only have
translations for French, and it's one more reason to finish migrating
the modal out of AngularJS (we cannot use Vite with AngularJS).

[0]
https://tuleap.net/plugins/git/tuleap/tuleap/stable?a=blob&hb=fcd7aaac3fd5392ead09280ad01a33096b2a5681&f=plugins%2Fprogram_management%2Fdocs%2Fdecisions%2F0005-build-domain-objects.md

Change-Id: Iec172ffbfdfd0816cd9944016ec4d739f1ea98b3