feat: ability to update artifact section via artidoc REST
Artidoc frontend should speak directly to artidoc backend instead of
speaking to tracker REST API.

This first contribution implements the update of the artifact section
when one is using PUT artidoc_sections/:id that already exists for
freetext sections.

Update of frontend will be done in a dedicated contribution.

Part of story #41805: Management of titles

Change-Id: Ica7ed5db2cbf35b22104eed224956e1b36df0d49