fix: request #41823 Remaining effort badge appears on every card
How to test:
In taskboard, not every card should have a remaining effort badge under
it. It should only appear when the card has a remaining effort field
value to a number or zero. When the field is empty, the badge should not
appear, the value in the JSON payload should be null.

Why?
Introduced by d239615add51210f6fa7523ae4c84421ebcfac45. The issue is due
to the too wide cast: the value can be null, for example when the field
is empty, or when the field is computed (without manual value).

Change-Id: I5e2b22e3c130fc1507c4c8fc7118cbacaedf690d