refactor: Switch Program Backlog app to Vue 3
Closes request #41457 Migrate plugin-program-management-backlog to Vue 3

How to test:
When browsing a Program's Backlog, no functional change expected.

Why?
I replaced calls to interpolate() by $gettext() as it supports
interpolation with vue3-gettext.
I removed the duplicated type definitions to only use the State and
ConfigurationState types.

I did not remove the empty `<div>`s in ToBePlanned and
ProgramIncrementList. I tried, and it breaks the layout, everything
becomes stacked in a row, instead of columns. There is something to be
done CSS-wise, but it can be later.

I did not change the computed access to the store in ToBePlanned because
TypeScript was complaining about a difference in types. It can be done
later and should still function.

Change-Id: I7a175636e6b841d9a136821c16e04a60e7998383