feat!: introduce 'import' in post representation
Currently there are two different payloads to create a section:
* to create an artifact section: `{ artifact: { id: 123 } }`
* to create a freetext section: `{ content: { title, description } }`

We would like to be able to create an artifact section the same way that
freetext sections are created (`content`), but we must keep `artifact`
to still be able to import an existing artifact.

Keeping `artifact` and `content` at the same level will be misleading
for the consumer of the API: which one should I choose to create an
artifact section?

In order to remove ambiguities, `artifact` is now under `import`
properties, so that we have `import` versus `content` which is less
confusing.
* to create an artifact section: `{ import: { artifact: { id: 123 } } }`
* to create a freetext section: `{ content: { title, description } }`

In the next contributions this will become:
* to import an artifact: `{ import: { artifact: { id: 123 } } }`
* to create an artifact section: `{ content: { title, description, type: 'artifact' } }`
* to create a freetext section: `{ content: { title, description, type: 'freetext' } }`

No functional changes except in the breaking change of the API.

Part of story #41805: Management of titles

Change-Id: Ibb44ce107f490f0474fa79d74cfb865bf992a8e2