Convert taskboardHeader into composition API
part of request #41443 convert taskboard into vue3

No functional changes on taskboard header

Why?
We cannot go through `useState` directly to get the features. It wraps
the state in a `readonly()`, but Vue 2 does not support readonly for
arrays. The only solution is to access the state from `useStore`
directly, to avoid the `readonly()` array. This solution is not intended
to stay as it is, it should be changed after the switch to Vue 3.

Change-Id: Iba15a040180b8a86ac0f971bbb7a80e7c0d43288