chore: Add more logs for LDAP account retrieval or creation
Part of request #41802 Add logs in Ldap plugin

How to test:
- enable LDAP authentication in local.inc
- in project administration, go to a User group and click on the "Set
LDAP group binding" button
- Type the name of your LDAP group in the autocompleter, and click on
that group in the list
- in ldap_syslog, you should be able to see the users of the LDAP group
that were found in Tuleap or that were created in Tuleap, or that were
not created due to an issue
- to test the case where the user account is neither found nor created,
edit the file at /etc/tuleap/plugins/ldap/etc/ldap.inc and change
$sys_ldap_eduid to a broken attribute name (for example
"zemployeeNumber"). In your logs, you should see that Tuleap could not
create the user account.

Change-Id: I5946d9ee00a3b0b8abfc337892b32796442dc6e7