feat: ability to create artifact section via artidoc REST
Artidoc frontend should speak directly to artidoc backend instead of
speaking to tracker REST API.

This contribution implements the creation of an artifact section
when one is using POST artidoc_sections/:id that already exists for
freetext sections.

Update of corresponding frontend will be done in a dedicated contribution.

Part of story #41805: Management of titles

Change-Id: Ie25385491676c6d5ac002d3201450b47fb6dcf49