fix: Enable or disable toolbar items
Part of story #41805: Management of titles

Why?
Before adding the new button for modifying section levels, we need to fix the toolbar's issues.

What has been fixed?
The list items or the link item were not disabled like the other toolbar items. They have their own enable/disable behavior. For example, the bullet list must be disabled when the numbered list is activated, and vice versa.
To correct this, I have added an `is_toolbar_disabled` property to these items to allow item activation calls to be ignored when the toolbar is disabled.

How to test:
-- Go to an Artidoc
-> All the items in the toolbar should be disabled

-- Click on a section description
-> The items should be enabled

-- Select text and click on bullet list button
-> Bullet list button should be activated and ordered list button should be disabled
-- Remove the bullet list
-> Both list buttons should be enabled

Change-Id: I2dc1fb0e73f5100be0639f6dac1484da32f0f559