fix: Tracker list emptyness is not checked when put
Part of story #41799 Have multiple queries per widget

This is checked when running the query but not when updating it. It has
been catch since initial search is currently broken

*Testing:*

Create a new widget and type a query with a from returning not tracker.
When saving you should get the same error as for search

Change-Id: I186b2d02f44ef9004218bd37507a5858a442e629