fix: allow our pre-receive Go/WASM Git signature checker to read quarantined packs
go-git has an issue were it does not look for packs in the quarantined
directory. This is problematic because it means we will not be able to read the objects
we need to access. It was fixed in our fork (will see to upstream a proper fix) and
we took advantage of the fix to only read the quarantined objects,
it means less data needs to be loaded so it is faster.

We give read access to the gitolite group to the quarantined directory.
This is needed because we execute the pre-receive hook as `codendiadm`.
Git sets 0700 permission to this directory even if you set `core.sharedRepository`
to `group`.

The memory and execution time limits have been raised in order to
execute relatively complex pre-receive hooks.

Part of request #41439 Validate commits and tags pushed into the Tuleap repository

Change-Id: I9ce1610a0fa976dc9e81dcb51019edd858d39c9f