feat: add new items to the toolbar at specific locations
Part of story #41805: Management of titles

Why?
To insert the [H] button in the toolbar. The [H] button is internal to Artidoc. It must be handled from Artidoc and not from the Toolbar lib. To do this, we need to be able to add elements to the toolbar from the outside.
This commit allows elements to be added to specific places in the toolbar.

It is the responsibility of the application using the toolbar to pass the element to be added, and the position in which to place it, in the toolbar.

All item groups are listed as exported constants. To select the position of the additional component, you must choose between "before" and "after" an item group name. The list of items is then reconstructed by inserting the additional items at their specific places.

How to test:
 -- Add `v-bind:additional_elements="[{ position: 'after', target_name: TEXT_STYLES_ITEMS_GROUP, item_element: 'hello' },]"` in `EditorToolbar.vue` on line 42.
 -> This should display 'hello' in the toolbar, at the indicated position.

 -- Try using several positons
 -- Try with several additional items:
 ```
v-bind:additional_elements="[
             { position: 'before', target_name: BASIC_TEXT_ITEMS_GROUP, item_element: 'hi' },
             { position: 'after', target_name: TEXT_STYLES_ITEMS_GROUP, item_element: 'hello' },
             { position: 'before', target_name: SCRIPTS_ITEMS_GROUP, item_element: 'bye' },
         ]"
 ```

Change-Id: I941a236dab5229f09b19b2b672494c68707445a2