fix: request #41854 Cross-tracker verification of field value existance is not accurate
Let's take 2 Trackers A and B. Both have a static list field:
- A have "Todo" value in it
- B have "To do" value in it

With this TQL query we have an error:

`SELECT @id FROM @tracker.name IN('A', 'B') WHERE list_field IN('Todo', 'To do')`

It is because for each Trackers they have at least one value of the IN
which is not present in their available values.

The idea of the fix to reverse the check: if they have at least one
value of the IN in their available values we are ok.

*Testing:*

The specified case should works. Other cases should not be impacted

Change-Id: I37567d9b01900930240c817e526eae82d8f8288f