feat: add button [H] to change section levels
Part of story #41805: Management of titles

The [H] button is added to the toolbar to manage section levels.
As the button is only useful to Artidoc, it must be defined and managed by Artidoc, then the html is sent to the toolbar, which displays it. For this reason, we make Hybrids component.

The [H] button is hidden behind the same feature flag as the freetext section, as they will be activated at the same time.

How to test:
-- Go to an artidoc
-> The [H] button should be displayed next to the Style elements and must be disabled

-- Click on a section description
-> The [H] button should still be disabled

-- Click on a section title
-> The [H] button should be enabled

-- Click on [H] button
-> The heading corresponding to the current level should be shown in gray

-- Change the level of the section
-> The cancel and save buttons should appear
-> The displayed level should change and all other numbers should be updated

-- Click on save button and reload the page
-> The displayed level should be the same

-- Change the level of the section and click on cancel button
-> The displayed level should be back to the previous level

Change-Id: Iacf5d8d9f1ae9dd0bb711ead331db267a5a6ae06