tests: e2e test for referenced artifacts in artidoc sections
Part of story #41805: Management of titles

This contribution adds a test case for the management of
cross-references in artidoc sections (freetext + artifact).

An artifact is created in teh setup phase, and a reference will be
injected into a freetext and and artidoc section. We now make sure the
two references have been extracted.

Change-Id: Ib7fcffe7cbeef0fc178951d7a1d920507957d4d1