fix: references extraction test in artidoc is flaky
Part of story #41805: Management of titles

What's the issue?

Cypress is a bit too fast. Therefore it does not wait the sections to be
in a stable state before trying to type some new content in their
descriptions. Hence, the text containing the reference was sometimes
inserted at the wrong place (either before the title, either before the
description).

The proposition to fix it is the following: When a section is being
saved or created, we wait until the element having the
`[data-test=section-edition]` attribute inside the current section
disappears. It means that the changes have been applied with what has
been received from the api.

How to test:
- in artidoc.cy.ts, comment the lines 205 to 224 (performing sections
  deletions), otherwise you will not be able to check that the
references have been properly inserted into the sections.

- run the test
--> In the freetext section, the cross reference have been inserted
after the original description (not in the title nor before the original
description)

--> Same expectation for the artifact section

DON'T FORGET TO UNCOMMENT THE LINES 205 TO 224 BEFORE MERGING THIS FIX!

Change-Id: I075f48ca35cd8ed492b17a55a758903a2e920018