fix: request #41831 Panic when a Tuleap Function allocate more initial memory than allowed
In the way we currently do the setup of the engine and module I did not
find a better way to identify this issue than matching the error
message. It is still better than getting a crash (especially since we
are doing over an FFI call...) and it is easy to adjust during Wasmtime
upgrades since the test will fail.

I have committed both the WAT and compiled WASM files. We could have
compiled the WAT when needed but it would require to also pull `wat2wasm`
toolkit which seems overkill for a small example like that.

Change-Id: I0e946de2eb053a390d2a86196d2d7625d6629199