fix: request #41866 Arrow on services opened in new tab is not displayed
Since b76665d11de5ce99b05eea858185936855e822df, the dao returns an int
for the column `is_in_new_tab` but the method `isOpenedInNewTab` do a
triple check `=== '1'` on it. As the type is not the same it always
returns false.

By testing with `=== 1` the check is fixed. I also add a (int) cast to
be sure and avoid future problems

*Testing:*

- create a service with 'Open in new tab' checked
- create a service without 'Open in new tab'

-> the first one should have an arrow at the right of its label
   the second one should have not the arrow

Change-Id: Ib3fcde084a4e77f4b243ddb8c2715193d158f797