feat: PDF templates should have default styles
If a new version of Tuleap fixes a bug or add a new functionality in PDF
templates styling, then everyone will have to 1) read the release note
2) manually adjust each of their own templates to match the update.

Before the PDF template usage is widespread, we should introduce the
notion of "default style", aka a block of style that is readonly and
subject to change by Tuleap releases. It does not completely remove the
need for customers to check that their templates are still ok after a
Tuleap release, but it should decrease the need and the size of
adjustments.

This is actually needed to deactivate the automatic numbering for
artidoc document that manages manually its own section numbering
according to sections level (see story #41805).

Closes request #41868: PDF templates should have default styles

Change-Id: I2f2e8865710a89354f877639a5227a1eea8dc567