fix: Boolean("false") in js is true
Part of story #41799 Have multiple queries per widget

To know if user is admin or not, we pass a data-attribute to the div
containing the widget app. If the user is not admin, its value was
"false" but in js `Boolean("false")` is considered as true.

Now we use directly a boolean value

Allow non-admin user to switch query by the way, the fix blocked it

Note: all data-attribute are converted to a single json one

*Testing:*

Non admin user should not be able to edit the widget

Change-Id: Ic40cc47a46687f097e099d239fc8a9c99f46d878