fix: Move the suppression of errors of the GitActions class outside the config file
This fixes the error encountered when Psalm in taint analysis mode. In
taint analysis mode the dead code detection is off so it considers the
ignore part of the config as useless.

Note the properties are actually used in the GitActions class, the code
is causing some troubles to Psalm. Not really worth to investigate, this
God class should be split/rewritten anyway.

Part of request #42175 Psalm: 5.26.1 -> 6.8.7

Change-Id: Ia95030c8599409cc9b9ce146b8f3055c25b66e7e