tests: Create an artifact with all fields whould be covered by e2e
Part of request #42184 Add end-to-end tests to tracker report

How to test:
make tests-e2e should succeed

Why?
After recent issues done by adding typehinting over tracker, we had several regressions.
It puts on highlight that we have a lack of tests on tracker, as it's the core of Tuleap we should assert that fields and report are usable, or at least regular usage never break.

Note:
This is the first in a series of tests that will test the tracker report.
This test creates an artifiact with all field types.

Open list fields and artifacts links fields are left aside for the moment:
 - open list is based on prototypemulitselect and we did not found a way to interact with cypress, I have some doubts about the possibility to test this field
 - artifact link will be done later after the work on New artifat link.

Change-Id: If02840096e12f09f89a2f33b0a65484708833d81