refactor: add elements at the beginning or end of an item group
Part of story #41805: Management of titles

Additional items were added before or after an item group, but were not in the same item group. It is now possible to add items at the beginning or end of an item group.
This refactor is necessary because the [H] button that will arrive in a future commit must be located at the end of the item group that manages styles.

How to test:
-- Add additional elements in `EditorToolbar.vue` at line 42:
```
v-bind:additional_elements="[
            {
                position: 'after',
                target_name: TEXT_STYLES_ITEMS_GROUP,
                item_element: 'bye',
            },
            {
                position: 'at_the_end',
                target_name: TEXT_STYLES_ITEMS_GROUP,
                item_element: 'hello',
            },
        ]"
```
-> 'hello' should be written in the same item group as text style.
-> 'bye' should be written after the text style item group in its own item group.

Change-Id: I57bb0d6bc3d1daaea3d64e4f44d9351553036a0e