chore: Story for link field without API
part of story #41466 Have new Artifact Links in Artifact View

How to test ?
- No functional change.
- When you run storybook in tlp-doc, there is a new folder that contains
the story for the Link field. It mounts an empty link field. You can
play with the props to configure it slightly (you will only see the
default selection of "is Child of" link type, or not).
- When you change the link type, in the "Actions" tab of the story
drawer, you can see each "change" event that is triggered by the custom
element.

Why?
It allows devs to test this field in isolation, but still completely
mounted in a real browser, which is much better than unit tests. Since
API does not respond, there is not much functional, but that will be
improved later.
It allowed me to spot that we had not included the CSS for
cross-reference badges in the lib.

Change-Id: I2e2c749034f371de3b27849381dbb3518223419b