refacto: Add a builder for CrossTrackerQuery
Part of story #41799 Have multiple queries per widget

It makes the tests more focused on what we want to test in
CrossTrackerQuery, 99% of the time, we do not need to have a real query
title or description.
Also, it will be less painful when we need to add a new attribute in
CrossTrackerQuery, it avoids to change 2629819681641 files.

No functional change expected, tests should still pass

Change-Id: I7f641be479f49bf75840432ef7f4c4ea2d045d9a