feat: Add preview modal when using suggested query
Part of story #41799 Have multiple queries per widget

Before changing current edited query with a suggested one, preview it in
a modal with a message telling the user that content will be replaced

Also added a e2e test on this modal to be sure it is well integrated in
the widget flow

*Testing:*

Create a new query and click on a suggested query -> a modal should show
with all details about the suggested query. If you click on 'Overwrite
current query' your content is replaced. Otherwise (click on cancel or
outside the modal), the modal disappear and nothing more happen

Change-Id: I65fd2d696dc3162773059c48eae047df85ca9d02