fix: numbering in artidoc PDF export
Automatic numbering of sections provided in PDF templates is fine for
general cases when there is only one level of sections.

In the case of Artidoc, the section levels require that we have to do a
manual numbering. Therefore starting from now the default CSS of PDF
templates provide a specific class:
`section-title-with-delegated-numbering`.

To test, try to print as pdf an artidoc document with sections at
different level: there should not be two different numbering for
sections in the generated document.

Technical note:
In the css file, the rule could have been written this way:
```css
.section-title {
   &::before {
     … // automtic numbering
   }
}

.section-title-with-delegated-numbering::before {
  // disable automatic numbering
}
```

However, templates that have been created before Tuleap 16.5 already
contains the `.section-title::before` definition in user style. This
means that in that case the final CSS would look like:
```css
/// default style

.section-title {
   &::before {
     … // automtic numbering
   }
}

.section-title-with-delegated-numbering::before {
  // disable automatic numbering
}

/// user style (copy of default before 16.5)

.section-title {
   &::before {
     … // automtic numbering
   }
}
```

Due to CSS specificity and order of rules, the rule
`.section-title::before` would take precedence over
`.section-title-with-delegated-numbering::before`.

Therefore in order to prevent that, we need to include a more specific
rule that have less chance to be broken by existing templates:

```css
.section-title {
   &::before {
     … // automtic numbering
   }

   &.section-title-with-delegated-numbering::before {
     // disable automatic numbering
   }
}
```

That way, in the generated CSS:
```css
/// default style

.section-title {
   &::before {
     … // automtic numbering
   }

   &.section-title-with-delegated-numbering::before {
     // disable automatic numbering
   }
}

/// user style (copy of default before 16.5)

.section-title {
   &::before {
     … // automtic numbering
   }
}
```

`.section-title.section-title-with-delegated-numbering::before` will be
used because its specificity is greater that `.section-title::before`.

Part of story #41805: Management of titles

Change-Id: I84006ea3ffd5f2b7eb9abf6350b4f72964fc5c02