chore: Story for Link field with fake backend
part of story #41466 Have new Artifact Links in Artifact View

How to test ?
- No functional change.
- When you run storybook in tlp-doc, in the "Tracker > Fields > Link
Field" story, the link field is no longer empty, it has existing links.
- You can search for an artifact in the list of possible parents (only
with "is Child of" type).
- You can search for an artifact in the history or with full-text search
(with all types except "is Child of"). You should not find the artifact
under edition itself in these lists.
- You can search an artifact by its ID (with all types). If you type
"647", you should not be able to select the artifact (because we cannot
link the artifact to itself).
- You can mark links for removal, add new links, remove new links. In
the "Actions" tab of the story drawer, you should see a "change" event
for each of these actions.

Why?
It allows devs to test this field in isolation, but still completely
mounted in a real browser, which is much better than unit tests. See the
ADR for the "Why" of Mock Service Worker. The artifact creation part of
the API does not respond, this will come later.

Change-Id: I62eebe88beb158e4a8cd75ef57be1cead32f6755