fix: [H] button should not be enabled when editor is not focused
Part of story #41805: Management of titles

When at least one section in the artidoc contains a cross-reference,
And the document has done loading
Then the [H] button should not be enabled.

This behavior exists because when the async-cross-reference has done
rendering itself, the view is updated a second time. The boolean
"is_first_update" is then false. Since the default selection is placed
inside the section's title, then the button is enabled by the
enable-or-disable-toolbar plugin.

The strategy is to remove this boolean and check whether the view has
the focus or not to enable/disable the toolbar or the [H] button.

How to test:
- Pick up an artidoc having some sections containing cross references
- Refresh the page
--> Everything in the toolbar is disabled

- Click the title of a section
--> The [H] button is enabled
--> The other buttons of the toolbar are disabled

- Click the description of a section
--> The [H] button is disabled
--> The orther buttons of the toolbar are enabled

Change-Id: Iad1db776d7f0f3d2b7d270d2164bf8eee868c6b1