fix: remplace initial pending section
When user opens an empty (and configured) document, then a pending
section is automatically displayed. If user want to insert a freetext
section, then this new section should replace the old, automatic, one.

This is only for the initial state of the document, nothing changes when
the document contains many sections.

Part of story #41805: Management of titles

Change-Id: I3aace01003f86c147f2173e552fcd84679c7f612