fixes request #42199 GraphOnTracker raise errors when user create graphs
How to reproduce:
 - add a grpah of each type
 - errors are logged

With this patch no error in logs

Note:
This patch only deals withtypehinting errors, on user interface logs are still semi broken (the detail of error is not displayed, for burndown it tries to display some crap).
I won't fix it now, if we want to improve this, it should be handled by a dediacted request.

Change-Id: I11c8993520b12387a895911ac50eae84225db0c2