fix: Do not require configuration changes for instances using an SMTP relay on 587/TCP
This is a follow up to d15b15281639480e727208a429f1e0c2dd5a0664 in order
to limit the cases where an admin have to do a manual action.

This change rename the configuration setting to activate SMTP implicit
TLS to something explicit to avoid confusion.
We automatically assume that SMTP relays using the port 587/TCP are
using STARTTLS and not implicit TLS as per the specification [0] so admins do
not have to reset the configuration for the standard case.

[0] https://datatracker.ietf.org/doc/html/rfc8314#section-3.3

Change-Id: Ide079a28e3436b08c67ba587431f09b79cb70d77