fix: Widget title should be the default title name (Cross Tracker Search) at new query creation
Part of story #41799 Have multiple queries per widget

How to test:
 - Go to the query creation menu
=> The widget title should be the default title of the widget
 - Create/Save a new query
=> You are redirected to the read query menu and the success feedback is
displayed.
=> The widget title is updated with the first query title
 - Choose another query
=> The widget title is updated

TL;DR: The contribution fix the success message not displayed at query
creation because it was cleared too soon. The widget listen now a dedicated
event and the widget title is now updated after query creation.

This contribution also fix (for free) the feedback message "not" displayed when a
query creation is successed. The feedback was not reeally displayed because the
success feedback was cleared too soon after the query creation because
of the call SWITCH_QUERY_EVENT.

We should also stop rely on the SWITCH_QUERY_EVENT at the title update
because this event is declared and removed in the ReadQuery component,
so, when we go to the creation query, the ReadQuery component is
unmounted and the associated event are removed and are not listened
anymore. This will also lead to dangling event.

An other issue was the tight coupling between the SWITCH_QUERY_EVENT and
update of the widget title, it means the title should change only when a
query is chosen, while we need to change the title when the menu
displayed change too.

Also, to avoid dangling event when the app or component is destroyed, a
method which remove the listener has benn added.

Change-Id: I806ecf8a5586100a57a842fd0d1a9da1b6bbea92