test: Do not use dynamic data-test attributes
part of request #42184 Add end-to-end tests to tracker report

How to test:
make tests-e2e should pass

Why?
Some time ago when writing e2e tests, we used "dynamic" data-test based
on the fields shortname. We should not do that because it makes it
impossible to link the production code to the test (try searching for
"data-test", without the value). We still know that maybe _some_ test is
using this, but we cannot know which one. It also makes the production
code "compute" things just for tests purpose. It also gives one more
entry for XSS that we have to carefully purify.
Since they have been used a lot, it's going to take patience killing
dynamic data-test attributes completely. In the meantime, we should not
use them and instead use completely static data-test attributes
(sometimes we have to add to the attribute to avoid name clash). The
dynamic part is the same as for the real user: the name of the label,
the name of the value in the options, etc.

Also, I made the project private and added ProjectMember in members.

Change-Id: I334d3a8bd66fbb5b49f69168847c65aa7bb00b8b