fix: request #42220 Adding a column in sorted table triggers an SQL error
How to test
-----------

Given a tracker report table sorted on a column
When user add a new column to the report
Then the following error is generated, resulting an empty/broken column

Implementation
--------------

The patch also embed a modification of legacy DataAccessObject in order
to capture the full stacktrace of errors when a SQL error happens. Given
most methods here are deprecated, I had to update the psalm baseline.

The patch would have welcomed tests but I don't really know how to add
one that is relevant.

Change-Id: I81fd99e3cb20b11871dddbafcd6e8fb5cc297f9f