fix: unable to set pending section level
What has been fixed:

1. When the document has loaded and is empty, a default pending section
   is added to the document. In this case, the click listener in charge
of disabling the toolbar buttons when a click happens outside the
toolbar or an editor was setup twice, causing side-effects. Hence, on
click on the [H] button, its popover open but the button was considered
disabled: clicking any title level did nothing. The solution to fix this
issue is to remove the event listener when the component is unmounted,
assuring that the handler is always executed once, and not twice.

2. The state of a newly created section was broken: When a pending
   section is replaced in the collection, its state was also
overwritten. Since this state is not reactive (only its attributes are),
and the sections are keyed by internal_id (which never change), then two
instances of the same state were exising in the document: the old one
still used by the components, and the new one never used because never
retrieved. Hence, if we try to update the level of a newly created
section, it was never displayed "in edit mode".

Note: a new ref has_title_level_been_changed has been temporarily
added to the sections states to avoid complex checks to know
whether the section should be in edit mode when its content has been
edited OR when the level of its title have changed. A refactoring is
planned to manage the section level change into the state rather than
the section itself.

How to test:
- Create a new artidoc
- Fill in the title and the description of the default pending section
- Change its level
--> Its level has been changed

- Save the section
- Change its level again
--> The section is displayed in edit mode

- Cancel
--> The section is no more displayed in edit mode

- Click the [H] button
--> The disabled item in the dropdown matches the current level of the
section, and not the previously selected level.

- Edit the description
- Change the section level
- Using the H button, set the section level back to its original value
--> The section is still in edit mode because the description has been edited.

- Cancel
--> The section's content and level have been reset back to their former values.

- Run the cypress tests
--> Tests should be successful

Change-Id: I605814d5f6f8bebc96ed2ddcaebadc569787aed9