fix: wrong toolbar buttons activation when the editor is focused
Part of story #41805: Management of titles

When editor has lost the focus and is focused back, the cursor
is hidden then displayed again BUT no view update is triggered. Since
the enable-or-disable-toolbar plugin reacts to view updates, it was
never notified. Hence, it was sometimes required to clic a second time
in the editor to enable/disable the toolbar buttons.

Now, the plugin also listen to the focus event and activates/deactivates
the right toolbar buttons depending on the cursor position.

Note: We cannot make it work with only the onClick event because we also
need to activate/deactivate the toolbar buttons when the cursor
enters/leaves the title or the description.

How to test:
- Have a section description containing a cross reference
- When the document has done loading
--> The toolbar buttons are all disabled

- Click in the section title
--> The H button is enabled

- Change the title level
- Save the section
- Click the title again
--> The H button is enabled once again (you did not have to click the
description or any other part of the editor).

- Click out of the toolbar and the editor
--> Everything is disabled in the toolbar

- Using the tab key, focus the editor
--> If the cursor was in the title, only the H button should be enabled
--> If the cursor was in the description, only the H button should be
disabled

- Using the arrows, jump from the title to the description back and
  forth
--> When the cursor is in the title, only the H button should be enabled
--> When the cursor is in the description, only the H button should be
disabled

Change-Id: If48716ae0bcfe7098161730c18190274696148ed