feat: Add 'default' tag in query choose dropdown
Part of story #41799 Have multiple queries per widget

If a query is the default one of the widget, it is now marked in the
choose query dropdown

Note: I use trim in the test because textContent now have a trailing
whitespace like this -> 'Get all Talbot '. It seems to be caused by the
possible span.

*Testing:*

Have a default query, it should be marked so.

Change-Id: I2b531eaaba74ae18a1b75673e0a8414059c7b92b