feat: Fullscreen edition of Artidoc ✨
When you work an a small screen like on a laptop, you may need more
vertical space while editing your document.

By removing the project header and the project sidebar, you are at ease
and can focus on the editing.

For now the option to switch to fullscreen is in the ellipsis menu of
the document. We will see later if we find a better place in the
toolbar, we must see first how it can fit with the global cancel/save
buttons that should also be put in the toolbar.

In order to remove navigation elements (project header, project sidebar)
requires us to switch one element to fullscreen: the artidoc container.
However this is not that simple, because the popover menus that are used
in artidoc are moved in the document.body, outside of the artidoc
container (this is because we have to deal with a z-index nightmare). If
we continue that path then everything outisde the artidoc container is
not displayed in fullscreen, meaning that user cannot manage links for
example.

A first attempt to handle this would be to move the popover menus inside
artidoc container instead of document.body. This works fine but brings a
major usuability issue: the artidoc is constantly scrolled at the top
everytime a popover is shown :( We didn't find why this is doing that
and how to workaround it.

This left us with the solution to switch the whole document.body to
fullscreen and manually hide navigation elements. This is not ideal
because we have to whitelist such elements, but we didn't find a better
solution.

Closes request #42250: Fullscreen edition of Artidoc

Change-Id: Ic6514d02fe8008d79d6ac7203f58339c89a56eb7