refactor: Deprecate Tracker::getParent()
part of request #42238 Deprecate Tracker::getParentUserCanView() and
Tracker::getParent()

How to test:
No functional change expected in
- Link field when creating a new User Story in a Team project. The
Program Features are still listed as possible parents.
- Link field when creating a new Task. The open User Stories are still
listed as possible parents.
- Link field in the artifact modal. Program Features or User Stories are
still listed as possible parents (through the API).

Why?
Checking the permissions is not included in ParentInHierarchyRetriever
because it is a low-level API and we do not always want to check them.
It is up to the caller to know when to do the check, in this case
PossibleParentsRetriever knows it must check the permissions, so it
checks by itself.

Change-Id: I8243f01f47e753e167d0dcb6ed7b45d2e4351304