fix: request #42241 Ctrl+k should open the link menu instead of the image menu
The key binding Ctrl+k has been associated to the wrong action. It
should open the link menu instead of the image menu to match ckeditor's
behavior.

For the moment, no key binding allow to open the image menu. Let's see
if the need is raised by the end users.

How to test:
- Click somewhere in an editor
- Hit Ctrl+k
--> The link menu is open

Change-Id: I1546885566286874a19825c745db3431df50bb11