fix: weird display of section being saved/refreshed 💄
When a section is being saved or refreshed, then a visual feedback is
displayed to inform the user. To do recent work on whitespace, the
feedback is weird, it looks like the bottom of the section is missing.

In order to keep the whitespace small and have a visual feedback, no the
whole section (including text content) has a background.

Part of story #41805: Management of titles

Change-Id: I14447b6c5b83ae6c3060784010f21c29284b2f2e